src/Lore.Web/Client/@common/dialogs/components/dialog/dialog.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { EventEmitter } from '@angular/core';
import { configureTestSuite } from 'lib/test-utils';

import { DialogComponent } from './dialog.component';

describe('[Core] DialogComponent', () => {
  let component: DialogComponent;
  let fixture: ComponentFixture<DialogComponent>;

  configureTestSuite();

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [DialogComponent]
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(DialogComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should have resize event emitter', () => {
    expect(component.resize).toBeTruthy();
    expect(component.resize instanceof EventEmitter).toBe(true);
  });
});
